Fix test/Interpreter/dynamic_replacement_opaque_result.swift
rdar://53029978

+++ test/Interpreter/Inputs/dynamic_replacement_opaque1.swift
@@ -10,13 +10,13 @@ extension Int: P {
 }
 
 @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
-func bar(_ x: Int) -> some P {
+func bar(_ x: Int64) -> some P {
   return x
 }
 
 struct Container {
   @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
-  func bar(_ x: Int) -> some P {
+  func bar(_ x: Int64) -> some P {
     return x
   }
